requirements.txt
pandas
pdfplumber
mysql-connector-python
